opti délai réponse f42772d4ecc30d7787164f65a4e96fed3452cc8f

+++ src/DW1000NgRTLS.hpp
@@ -132,7 +132,7 @@ namespace DW1000NgRTLS {
     void transmitRangingConfirm(byte tag_short_address[], byte next_anchor[]);
     void transmitRangingConfirmExtended(byte tag_short_address[], uint32_t timePollReceived, uint32_t timeResponseToPoll, uint32_t timeFinalMessageReceive, byte net_id[], byte personal_short_address[]);
     void transmitRangingConfirmSingleDelayed(byte tag_short_address[], uint32_t timePollReceived, byte net_id[], byte personal_short_address[]);
-    void transmitRangingConfirmSingleDelayedSimple(uint32_t timePollReceived);
+    void transmitRangingConfirmSingleDelayedSimple(uint64_t timePollReceived);
     void transmitActivityFinished(byte tag_short_address[], byte blink_rate[]);
     
     boolean receiveFrame();
